Added unit test cases

+++ employee-system-api/src/main/java/com/example/demo/entity/EmployeeEntity.java
@@ -22,7 +22,7 @@ public class EmployeeEntity {
     
     @Column(unique = false)
     private String empName;
-    private long salary;
+    private Double salary;
     
     
     @ManyToOne(fetch = FetchType.LAZY)
